Change project hub page to static HTML, use templating for wireframes
The wireframe page is really just a list of links and uses very little templating. The wireframe pages share a lot of markup (navigation, footer, related posts section) so it makes sense to use the templating engine for this instead. Moved repeated markup into _includes files and referenced in _wireframes files.